refactor: extract CreateGenesisBlock, move LoadBlockIndex to blockstorage
Extract the genesis-block construction (timestamp coinbase, regtest
premine, header fields, dev-mode nonce search, hash/merkle asserts) out
of LoadBlockIndex into a CreateGenesisBlock() helper in chainparams.cpp,
fulfilling the old "uncomment this when CBlock is moved from main.h"
TODO that sat there (CBlock moved to primitives/block.h in #3060) --
though as a fresh transplant of the live code rather than the bit-rotted
commented-out draft, which predates the regtest premine and would have
changed the testnet coinbase nTime. The genesis hash constants move from
main.h to chainparams.h beside it.

The remaining LoadBlockIndex shell (network-mode settings, LevelDB
index load, genesis write-to-disk + AddToBlockIndex + regtest tx-index
seed, chain-trust/seen-stakes priming) moves to node/blockstorage.{h,cpp}
with the other block-file I/O (issue #3125, workstream C5).

Consumers are unaffected: main.h already includes chainparams.h and
node/blockstorage.h, so both the constants and the LoadBlockIndex
declaration still resolve transitively. Behavior-preserving; the only
non-move edit is std:: qualification of one vector in the transplanted
code (chainparams.cpp has no using-directive).

main.cpp -215 lines.
