+++ src/node/blockstorage.h
@@ -30,6 +30,14 @@ FILE* AppendBlockFile(unsigned int& nFileRet);
 bool LoadExternalBlockFile(FILE* fileIn, size_t file_size = 0,
                            unsigned int percent_start = 0, unsigned int percent_end = 100);
 
+//! Load the block index from LevelDB, creating and writing the genesis block
+//! first on an empty datadir (via CreateGenesisBlock in chainparams.cpp).
+//! Moved from main.{h,cpp} (issue #3125, workstream C5). Takes cs_main
+//! internally; callers MUST NOT hold cs_main when calling (the internal LOCK
+//! would deadlock under non-recursive locking; cs_main is currently recursive
+//! but the annotation contract documents the intent).
+bool LoadBlockIndex(bool fAllowNew=true);
+
 
 #endif // BITCOIN_NODE_BLOCKSTORAGE_H
